Move DatabaseSchemas into core/
